refactor: Feature Based Patternにディレクトリ構成を変えるためにPlayHistory, ToggleOrderを分離
