+++ src/TicTacToe/GameStatus/models/model.ts
