+++ src/TicTacToe/models/model.ts
